requirements.txt
pydantic>=2.0.0
litellm>=1.0.0
pypdf>=3.0.0
python-docx>=0.8.11
pillow>=10.0.0
requests>=2.31.0
tqdm>=4.66.0
openai>=1.0.0
huggingface_hub>=0.17.0
diffusers>=0.21.0
transformers>=4.33.0
accelerate>=0.23.0
safetensors>=0.4.0
